Minor fix to avoid warnings.
git-svn-id: file:///home/behr_mi/git/sumo_synched/branches/parking@21907 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/router/ROLoader.cpp
@@ -208,7 +208,7 @@ ROLoader::processRoutes(const SUMOTime start, const SUMOTime end, const SUMOTime
     SUMOTime lastStep = firstStep;
     SUMOTime time = MIN2(firstStep, end);
     while (time <= end) {
-        writeStats(time, start, absNo, endGiven);
+        writeStats(time, start, (int)absNo, endGiven);
         myLoaders.loadNext(time);
         if (!net.furtherStored() || MsgHandler::getErrorInstance()->wasInformed()) {
             break;
